update integration test for incoming call
