+++ packages/ringcentral-widgets-test/test/integration-test/IncomingCallPage/index.spec.js
@@ -1,6 +1,9 @@
+import CircleButton from 'ringcentral-widgets/components/CircleButton';
+import ActiveCallPad from 'ringcentral-widgets/components/ActiveCallPad';
 import IncomingCallPad from 'ringcentral-widgets/components/IncomingCallPad';
-import { inboundSession } from '../../support/session';
+import ActiveCallButton from 'ringcentral-widgets/components/ActiveCallButton';
 import { getWrapper } from '../shared';
+import { getInboundCall } from '../../support/callHelper';
 
 let wrapper = null;
 let phone = null;
@@ -18,14 +21,32 @@ beforeEach(async () => {
   });
 });
 
-async function getInboundCall(session = inboundSession) {
-  await phone.webphone._webphone.userAgent.trigger('invite', session);
+async function makeInbountCall() {
+  await getInboundCall(phone, {
+    id: '111',
+    direction: 'Inbound',
+    _header_callId: 'call-111'
+  });
   wrapper.update();
 }
 
 describe('Incoming Call Interaction', () => {
   test('When user has an incoming call, page should display Incoming Call Page', async () => {
-    await getInboundCall();
+    await makeInbountCall();
     expect(wrapper.find(IncomingCallPad)).toHaveLength(1);
   });
 });
+
+describe('Inbound Call in Call Control Page', () => {
+  test('RCI-1038#2 - User anwser the incoming call, Add button is disabled in Call Control Page', async () => {
+    await makeInbountCall();
+    const buttonAnswer = wrapper.find(IncomingCallPad).find(ActiveCallButton).at(4);
+    buttonAnswer.find(CircleButton).simulate('click');
+    wrapper.update();
+    expect(phone.routerInteraction.currentPath).toEqual('/calls/active');
+    const buttonAdd = wrapper.find(ActiveCallPad).find(ActiveCallButton).at(3);
+    expect(buttonAdd.find('.buttonTitle').text()).toEqual('Add');
+    expect(buttonAdd.props().disabled).toBeTruthy();
+  });
+});
+
